Return an error code if yum can't install any of the specified packages
 People want a bash scriptable way to say "X should be installed". So
with this patch:

. If _any_ package specified: is installed already; or is found and will
be installed == yum returns 0 (or at least goes into depchecking etc.)

. If _all_ package(s) are "not found", yum turns the message "Nothing to do"
into an error (returns 1, atm).

...thus:

1. yum install -y a b && echo worked

This will echo "worked" if _either_ a or b is installed after yum
finishes, but tries to install both.

2. yum install a && yum install b && echo worked

This will echo "worked" if both are installed (and if a is not
available, b will not even try to be installed).